lesson 4

+++ src/lesson4/task1/List.kt
@@ -240,10 +240,10 @@ fun factorizeToString(n: Int): String {
  */
 fun convert(n: Int, base: Int): List<Int> {
     var num = n
-    var b = 1
+    var b = 1L
     val countBase = mutableListOf<Int>()
     while (b <= n) {
-        countBase.add(b)
+        countBase.add(b.toInt())
         b *= base
     }
     val result = mutableListOf<Int>()
